UniSolve-Backend-main/requirements.txt
Flask
flask-mysqldb
flask-cors
bcrypt
PyJWT
pytz
requests
mysql-connector-python
pymysql
